requirements.txt
Flask
python-dotenv
azure-cosmos
gunicorn